requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai>=0.5.0
pydantic
python-multipart